+++ library/hls/src/main/java/com/google/android/exoplayer2/source/hls/playlist/HlsMediaPlaylist.java
@@ -393,9 +393,12 @@ public final class HlsMediaPlaylist extends HlsPlaylist {
    */
   @PlaylistType public final int playlistType;
   /**
-   * The start offset in microseconds, as defined by #EXT-X-START.
+   * The start offset in microseconds, as defined by #EXT-X-START, or {@link C#TIME_UNSET} if
+   * undefined.
    */
   public final long startOffsetUs;
+  /** Whether the start position should be precise, as defined by #EXT-X-START. */
+  public final boolean preciseStart;
   /**
    * If {@link #hasProgramDateTime} is true, contains the datetime as microseconds since epoch.
    * Otherwise, contains the aggregated duration of removed segments up to this snapshot of the
@@ -480,6 +483,7 @@ public final class HlsMediaPlaylist extends HlsPlaylist {
       String baseUri,
       List<String> tags,
       long startOffsetUs,
+      boolean preciseStart,
       long startTimeUs,
       boolean hasDiscontinuitySequence,
       int discontinuitySequence,
@@ -498,6 +502,7 @@ public final class HlsMediaPlaylist extends HlsPlaylist {
     super(baseUri, tags, hasIndependentSegments);
     this.playlistType = playlistType;
     this.startTimeUs = startTimeUs;
+    this.preciseStart = preciseStart;
     this.hasDiscontinuitySequence = hasDiscontinuitySequence;
     this.discontinuitySequence = discontinuitySequence;
     this.mediaSequence = mediaSequence;
@@ -575,6 +580,7 @@ public final class HlsMediaPlaylist extends HlsPlaylist {
         baseUri,
         tags,
         startOffsetUs,
+        preciseStart,
         startTimeUs,
         /* hasDiscontinuitySequence= */ true,
         discontinuitySequence,
@@ -605,6 +611,7 @@ public final class HlsMediaPlaylist extends HlsPlaylist {
         baseUri,
         tags,
         startOffsetUs,
+        preciseStart,
         startTimeUs,
         hasDiscontinuitySequence,
         discontinuitySequence,